AP_NavEKF3: use external class for EKF-GSF
