+++ libraries/AP_NavEKF3/AP_NavEKF3_Outputs.cpp
@@ -637,3 +637,7 @@ void NavEKF3_core::getOutputTrackingError(Vector3f &error) const
     error = outputTrackError;
 }
 
+void NavEKF3_core::getDataEKFGSF(float *yaw_composite, float *yaw_composite_variance, float yaw[N_MODELS_EKFGSF], float innov_VN[N_MODELS_EKFGSF], float innov_VE[N_MODELS_EKFGSF], float weight[N_MODELS_EKFGSF])
+{
+    yawEstimator.getLogData(yaw_composite, yaw_composite_variance, yaw, innov_VN, innov_VE, weight);
+}
