+++ libraries/AP_NavEKF3/AP_NavEKF3_VehicleStatus.cpp
@@ -393,10 +393,6 @@ void NavEKF3_core::detectFlight()
 
     }
 
-    // store current on-ground  and in-air status for next time
-    prevOnGround = onGround;
-    prevInFlight = inFlight;
-
     // Store vehicle height and range prior to takeoff for use in post takeoff checks
     if (onGround) {
         // store vertical position at start of flight to use as a reference for ground relative checks
@@ -411,6 +407,21 @@ void NavEKF3_core::detectFlight()
             yawInnovAtLastMagReset = innovYaw;
         }
     }
+
+    // handle reset of counters used to control how many times we will try to reset the yaw to the EKF-GSF value per flight
+    if (!prevOnGround && onGround) {
+        // landed so disable filter bank
+        EKFGSF_run_filterbank = false;
+    } else if (!prevInFlight && inFlight) {
+        // started flying so reset counters and enable filter bank
+        EKFGSF_yaw_reset_time_ms = 0;
+        EKFGSF_yaw_reset_count = 0;
+        EKFGSF_run_filterbank = true;
+    }
+
+    // store current on-ground  and in-air status for next time
+    prevOnGround = onGround;
+    prevInFlight = inFlight;
 
 }
